about halfway done with pytwitter conversion
OK, so seriously, git stash is the worst, which
is the whole reason this crappy commit
exists in the first place. Anyway...

There are tests for the two 'parsing' functions,
but need event data before the rest of it can
be tested properly.

Using pytest required the addition of a few files.
Specifically a setup.py, a few __init__s, pytest.ini
etc. It's all boring packaging stuff.

Additionally, requirements have been properly
broken out into production and testing and I
fixed an issue I didn't even know I had where
some of the Greek Monster names had a space
in them, which cause the GIF to be unreachable
from the web (and broke sending the GIF DM too.)